1.00

+++ src/user/user.controller.ts
@@ -17,6 +17,9 @@ import { User } from './schemas/user.schema';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { JwtAuthGuard } from 'src/auth/guards/jwt.auth.guard';
+import { Roles } from 'src/auth/guards/roles.decorator';
+import { RolesGuard } from 'src/auth/guards/roles.guard';
+import { UpdateContactDto } from './dto/update-contact.dto';
 
 @Controller('user')
 export class UserController {
@@ -28,54 +31,54 @@ export class UserController {
   }
 
   @Get()
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  @Roles('isAdmin')
   async findAll(): Promise<User[]> {
     return this.userService.findAll();
   }
 
   @Get(':id')
-  async findOne(@Param('id') id: string): Promise<User> {
+  @UseGuards(JwtAuthGuard)
+  async findOne(@Param('id') id: string, @Request() req): Promise<User> {
     const user = await this.userService.findOne(id);
     if (!user) {
       throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
     }
+
+    const requester = req.user;
+
+    if (requester._id !== id && !requester.isAdmin) {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar este usuário',
+      );
+    }
+
     return user;
   }
 
   @Patch(':id/set-admin')
-  @UseGuards(JwtAuthGuard)
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  @Roles('isAdmin')
   async setAdminStatus(
     @Param('id') id: string,
     @Body('isAdmin') isAdmin: boolean,
-    @Req() req: any,
   ): Promise<User> {
-    // Só um admin pode promover outro
-    if (!req.user.isAdmin) {
-      throw new NotFoundException(
-        'Acesso negado: apenas administradores podem promover usuários.',
-      );
-    }
-
     return this.userService.setAdminStatus(id, isAdmin);
   }
 
-  @Patch(':id/promote')
-  @UseGuards(JwtAuthGuard)
-  async promoteToProfessor(
+  @Patch(':id/set-professor')
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  @Roles('isAdmin')
+  async setProfessorStatus(
     @Param('id') id: string,
-    @Request() req,
+    @Body('professor') professor: boolean,
   ): Promise<User> {
-    const user = req.user;
-
-    if (!user.isAdmin) {
-      throw new ForbiddenException(
-        'Acesso negado. Apenas administradores podem fazer isso.',
-      );
-    }
-
-    return this.userService.update(id, { professor: true });
+    return this.userService.setProfessorStatus(id, professor);
   }
 
   @Patch(':id')
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  @Roles('isAdmin')
   async update(
     @Param('id') id: string,
     @Body() updateUserDto: UpdateUserDto,
@@ -87,12 +90,41 @@ export class UserController {
     return updatedUser;
   }
 
+  @Patch(':id/contato')
+  @UseGuards(JwtAuthGuard)
+  async updateContact(
+    @Param('id') id: string,
+    @Body() dto: UpdateContactDto,
+    @Request() req,
+  ): Promise<User> {
+    const requester = req.user;
+
+    if (requester.sub !== id && !requester.isAdmin) {
+      throw new ForbiddenException('Sem permissão para alterar este contato');
+    }
+
+    return this.userService.update(id, dto);
+  }
+
   @Delete(':id')
-  async remove(@Param('id') id: string): Promise<{ message: string }> {
+  @UseGuards(JwtAuthGuard)
+  async remove(
+    @Param('id') id: string,
+    @Request() req,
+  ): Promise<{ message: string }> {
+    const requester = req.user;
+
+    if (requester.sub !== id && !requester.isAdmin) {
+      throw new ForbiddenException(
+        'Você não tem permissão para deletar este usuário',
+      );
+    }
+
     const result = await this.userService.remove(id);
     if (!result) {
       throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
     }
+
     return { message: `Usuário com ID ${id} deletado com sucesso` };
   }
 }
